// utils/enums.tsx
export enum Animals{
  cachorro = "Cachorro",
  cavalo = "Cavalo",
  galinha = "Galinha",
  gato = "Gato",
  porco = "Porco",
  outro = "Outro"
}

export enum ColorsOptions {
  black = "black",
  blue = "blue",
  gray = "gray",
  green = "green",
  purple = "purple",
  red = "red",
  white = "white",
  yellow = "yellow",
}

export enum InputModeOptions {
  decimal = "decimal",
  email = "email",
  none = "none",
  numeric = "numeric",
  search = "search",
  text = "text",
  tel = "tel",
  url = "url",
}

export enum KeyboardTypeOptions {
  decimalpad = "decimal-pad",
  default = "default",
  emailaddress = "email-address",
  numberpad = "number-pad",
  numeric = "numeric",
  phonepad = "phone-pad",
  url = "url",
}